refactor: extract account methods
