+++ src/main/java/com/mandacarubroker/controller/UserController.java
@@ -5,7 +5,6 @@ import com.mandacarubroker.domain.user.RequestUserDTO;
 import com.mandacarubroker.domain.user.ResponseUserDTO;
 import com.mandacarubroker.service.UserService;
 import jakarta.validation.Valid;
-import jakarta.validation.constraints.Positive;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RestController;
@@ -15,7 +14,6 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.DeleteMapping;
 
 import java.util.List;
@@ -63,18 +61,6 @@ public class UserController {
         }
 
         return ResponseEntity.ok(updatedUser.get());
-    }
-
-    @GetMapping("/deposit")
-    public ResponseEntity<ResponseUserDTO> deposit(@RequestParam @Positive final double amount) {
-        ResponseUserDTO user = userService.doDepositForAuthenticatedUser(amount);
-        return ResponseEntity.ok(user);
-    }
-
-    @GetMapping("/withdraw")
-    public ResponseEntity<ResponseUserDTO> withdraw(@RequestParam @Positive final double amount) {
-        ResponseUserDTO user = userService.doWithdrawForAuthenticatedUser(amount);
-        return ResponseEntity.ok(user);
     }
 
     @DeleteMapping("/{id}")
